+++ tools/build-armor-compendium.mjs
@@ -44,7 +44,7 @@ const cleanStats = {
   exportSource: null,
   coreVersion: "13.351",
   systemId: "swnr",
-  systemVersion: "2.3.0"
+  systemVersion: "2.3.1"
 };
 
 await fs.rm(sourcePack, { recursive: true, force: true });